chor(elation): move parser to dedicated file for consistency
