+++ extensions/elation/actions/updatePatientTags/lib/getTagsFromLLM/parser.ts
@@ -1,5 +1,24 @@
+import { z } from 'zod'
 import { StructuredOutputParser } from '@langchain/core/output_parsers'
-import { TagsOutputSchema, type TagsFromAI } from '../../config/types'
+
+export const SingleTagSchema = z.string()
+
+export const TagsSchema = z
+  .array(SingleTagSchema)
+  .describe('The updated array of tags')
+
+export const TagsOutputSchema = z.object({
+  updatedTags: TagsSchema,
+  explanation: z
+    .string()
+    .describe('A single, concise sentence summarizing all tag modifications, including the reasoning behind the changes or lack thereof.'),
+})
 
 export const parser = StructuredOutputParser.fromZodSchema(TagsOutputSchema)
-export type { TagsFromAI } 
+
+export type TagsOutput = z.infer<typeof TagsOutputSchema>
+
+export interface TagsFromAI {
+  validatedTags: string[]
+  explanation: string
+}
